Fix for bug #1100
Under Windows, Quicktime images with a depth < 32 didn't
show up in the sequence editor. To fix this a non transparent layer was added to them. (bf227-bf232)

I now have disabled this trick for GIF images, so the transparency is preserved. But this is a quick fix and
should be replaced by a transparency check function.