Popular request: the floating panels (NKEY) now close on a 2nd NKEY command. This for all current NKEY menus by default. Can be standard for each hotkey controlled Panel from now on; use toggle_blockhandler() for it.
Another popular request: hotkey enabled Panels now optionally reopen on
the previous location, instead of under mouse. This option "Pinned" is a
user option now. Optional for later; have it as per-panel option... dunno!

The last request, close on mouse-exit, won't be implemented, this intrudes
with Panel consistancy, and is pretty OK solved with NKEY-NKEY now.