<sob>dropping OpenGL float buffer drawing... it seems that this isn't a well supported feature for glDrawPixels(), especially on ATIs crashes happen with it irregularly. (note; drawing float is a regular opengl 1.0 feature...).
Now all scanline/part updates in renderwindow is converted to 32 bits
before drawing, also the in the UV Image window only 32 bits RGBA is drawn.

What is still float: drawing the final image after render in renderwindow,
and drawing the Z or Alpha in renderwindow and UV image.

Of course we need to test this first. :) Will await reports...