Bugreport #3891
Object Layer Ipos didn't work when the Ipo was moved an Object Action.
Can't get this to work though... this option has been added with a lot
of exception handling already, and for action/nla it's very nasty to
blend/mix/add layer bit values. So; better not support that.

As feedback for users, I've added a notice popup when you try to move
layer-ipos to an action.