I Added a dummy Error Handler temporarly while it trys to open the tablet and friends up....
Basically the issue this fixes is this.  On my linux machine at
work someone setup a generic xorg.conf file that defines all
kinds of devices (a tablet and other stuff) which are not on my
machine.  So blender was exiting when it tried to open the tablet
and failed.  Now it doesn't.... :)

I also did some very minor reformatting... (replaced parts that had spaces
with tabs)

Kent