A possible problem was noted by Cyril Brulebois regarding copyright issues with regards to the noise functions in yafray based on Ken Musgrave's original code. I had left the 'Copyright' notice in the comments, and according to Cyril Brulebois this is a problem. In fact, from what I understand this makes it even impossible to use or modify the code in other software. But since it is not a verbatim copy of the code but rather based on Musgrave's work, he suggested I change it to explicitely state that it is in fact based on the code from the 'Texturing & Modeling' book. And since the yafray code is in turn based on the blender code, I better adapt the blender code too.
This reminded me that I also have forgotten to include the copyright notice
in the mersenne twister rng code I used for the Python Noise module.
This does clearly state to include the original notice with any resdistributed
code, in modified form or not. So I added that too.

I hope that solves the problems.