=IDProperties Python update=
Added a test gui script for ID Properties.  To use,
just select ID Property Browser from the help menu.
You can browse, create and edit ID Properties
for several library block types.  Should be
fairly straightforward to use.

This is a test gui script; once an actual ID property
gui is implemented in blender it will be removed.  It
is *not* going to be in the release, obviously, so once
the real gui is implemented this script will be removed.
Also, this being a testing non-release script it seems to me
to be okay to add it without all the proper review stuff.  If
not, I can remove it.

I believe a test script is really, really needed to properly
test/review id properties, so I feel this is worth it.