Bugfix #5625:
When adding keyframes to a very-short action scaled to long-lengths in the
NLA editor, exisiting keyframes would disappear. The culprit was a very large
threshold value used to cull duplicate keyframes. This bug was even present
in 2.42(a).

Although this fixes the bug, further work is needed on preventing future cases
of badly scaled actions. Actions with length of 0.9 frames representing motion
of 200 frame ranges is not on.