== IPO Cleaning ==
Rewrote the core-function responsible for ipo-cleaning. Now, it is in a less
wacko form. What is still not done is the conversion of a bunch of points
describing an arc to an arc defined by the handles of the keyframes on
either side of the arc. That will have to wait for the next development cycle.

This rewrite fixes two big bugs with the code:
* All but first curve got hidden aften cleaning
* Cleaning a curve with only two verts resulted in only one vert, even though
the values were not the same.