Bugfix: 'Random' crashes after duplicating bones
This commit should fix some seemingly random crashes broken and I have been experiencing while editing armatures.

A backtrace revealed that autosave was choking on the PoseChannels that didn't have a Bone assigned to them. This was caused by the bone duplication code making a new PoseChannel for a duplicated bone, but that new bone not getting assigned to the PoseChannel yet, as the user was still in EditMode.