transform mode  commit
you can now translate or rotate a selected object
in the third method. either in object or edit mode.

Note : the method is not ideal, as we create a full
transform loop for each component of the transform.
However it is this way handled in same way as other
means of input