Axial Symmetry Restoration
Mirror merge two sides together to restore axial symmetry (like on a humanoid).

The symmetry is detected using the topological information from the Reeb graph. Eventually, it will have to distinguish between topological symmetry and potential spatial symmetry and only restore the former (for example, a human is always topologically symmetric, but we only want to restore the symmetry if the limbs are oriented in the same way).