Skeleton Sketching
Cut gesture: drawing a line with RMB adds exact points (joints in the final skeleton) on strokes. Can cut multiple strokes at once as long as it cuts each one only once (this could be changed).

Shape of the gesture doesn't matter, but eventually it would be straight lines shapes for cuts.