missed this in last commit comments, Pattern Select operator, access from the object select menu
Glob strings like "Lear*.brown" and "Tree.0?", option for case sensitive and extend the existing selection.

currently the default string is "*" which needs to be edited in the redo-panel in the toolbox since there is no way to get a text input for python.
This replaces 2.4x's data browser Shift+F4, pattern select.