Bugfix [#25221] Jump to Next/Prev keyframe doesn't work with RotoBeizer
The jump to keyframes operator was being a bit too strict with its
checks for where it should check for keyframes.

In this particular case, RotoBezier keyframes were on Curve AnimData,
so the check for Object AnimData would return false, thus overlooking
this possibility. However, it should be safe enough to just open
things up a bit more.