Consolidation of stroke drawing.
Fixed a complicated bug that caused a failure of CurvePoint::getFEdge()
which had affected a number of C/Python API functions such as MaterialF0D.

The current view map building procedure may generate ViewEdges whose
two-dimensional (2D) length is almost or exactly zero.  Such a zero-length
ViewEdge is possibly chained with other ViewEdges to form a stroke.  When
the stroke is finally generated by Operators::create(), an attempt to remove
redundant vertices at the same 2D point is made.  This possibly breaks the
links of ViewEdges on top of which the stroke has been built, and eventually
result in a fatal error of CurvePoint::getFEdge() when API functions that
rely on this method are called from within a style module.

The present fix addresses this issue by automatically removing zero-length
ViewEdges (and Chains of them) before stroke drawing is started and after
splitting is performed (e.g., using Operators::sequentialSplit()).