Fix for a crash in silhouette edge detection.  The problem was reported by macouno.  Thanks!
The crash was caused by a lack of curvature information required
for smooth edges.  Now the curvature information is computed if and
only if there are smooth edges.  This leads to a minor performance
improvement, because in the past the curvature information was
always computed when the Face Smoothness was enabled.

(To be precise, the above description is true when both the Ridges
and Valleys and Suggestive Contours options are disabled.  If they
are enabled, the curvature information is always computed because
it is necessary for the determination of these edge natures.)