NLA Drawing - When "Show Control Curves" option in View menu is disabled, the strips are drawn so that they take up less vertical space.
Originally, the primary reason why these were taller than those in the
other animation editors was really so that these control curves could
be visualised adequately. So, when these aren't shown, we can afford
to collapse the strips vertically.

This should make it possible to fit more strips on screen to retime
them. in some staggered fashion.