NLA Drawing - Second attempt at providing options for streamlining the view for transforming strips
When the "Include animation data blocks with no NLA data" toggle
(action icon) is off, action lines are only shown if they have
keyframes. So when this option is off, only NLA blocks that have NLA
tracks will be shown, and of those, only those which currently have an
active action with keyframes will have their red action lines shown.

Combined with the vertical-space tweak when show control curves is
turned off, this should be good enough for most cases.