Bugfix #33161
The new node that outputs multilayer was using longer names than default.
Caused old code that truncated pass names to 11 chars to fail on loading exr.

This was an old limit in openexr - but that got fixed long ago.

On todo: check current openexr name lenghts, and all code in Blender that
defines pass/layer names.